Refactor: Modularize Streamlit application
I decomposed the main `st_app.py` into several focused modules:
- `gcs_utils.py`: Handles Google Cloud Storage interactions.
- `api_client.py`: Manages fetching data from the external API.
- `data_processing.py`: Contains logic for loading and processing restaurant data.
- `bq_utils.py`: Handles BigQuery operations, including column sanitization and data writing.

I updated `st_app.py` to delegate functionality to these new modules, significantly reducing its size and improving readability.

I modified `test_st_app.py` to align with the new modular structure, ensuring that all tests pass by updating import paths and mock targets. The existing test coverage is maintained.

This refactoring enhances code organization, maintainability, and reusability without altering the application's functionality.
